+++ activity-platform/frontend-utilities/docs/convert-to-error.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { MarkdownRenderer } from '../src/documentation/markdown-renderer';
+// @ts-ignore
+import input from '../src/error-handling/convert-to-error/README.md';
+
+export default <MarkdownRenderer>{input}</MarkdownRenderer>;