+++ activity-platform/frontend-utilities/docs/retry-operation.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { MarkdownRenderer } from '../src/documentation/markdown-renderer';
+// @ts-ignore
+import input from '../src/network/retry-operation/README.md';
+
+export default <MarkdownRenderer>{input}</MarkdownRenderer>;