+++ activity-platform/frontend-utilities/src/index.ts
@@ -1,3 +1,4 @@
+export { MarkdownRenderer } from './documentation/markdown-renderer';
 export { convertToError } from './error-handling/convert-to-error';
 export { FailedFetchError, retryOnException } from './network/retry-operation';
 export { STORAGE_MOCK, mockWindowStorage } from './storage/local-storage';