+++ linking-platform/link-picker/examples/10-with-plugins.tsx
@@ -6,14 +6,15 @@ import {
   AtlassianLinkPickerPlugin,
   Scope,
 } from '@atlassian/link-picker-atlassian-plugin';
-import { mockRecentClient } from '../example-helpers/providers';
+import { mockEndpoints } from '@atlassian/recent-work-client/mocks';
+import mockRecentData from '../example-helpers/mock-data';
 import { LinkPicker, LinkPickerPlugin } from '../src';
 
 type OnSubmitPayload = Parameters<
   Required<React.ComponentProps<typeof LinkPicker>>['onSubmit']
 >[0];
 
-mockRecentClient();
+mockEndpoints(undefined, undefined, mockRecentData);
 
 export default function InPopup() {
   const [isOpen, setIsOpen] = useState(true);