+++ linking-platform/link-picker/src/common/constants.ts
@@ -1,0 +1,3 @@
+export const ANALYTICS_CHANNEL = 'media';
+
+export const COMPONENT_NAME = 'LinkPicker';